Refactor and optimize goods and cart management for better performance and maintainability

+++ miniprogram/package-goods/services/goods/create-goods-spu.js
@@ -1,32 +1,36 @@
 import log from '@/common/log/log';
-import { saasId } from '@/common/saas/saas';
 import goodsModel from '../../models/goods/index';
 import goodsStore from '../../stores/goods/index';
-import { runInAction } from 'mobx-miniprogram';
 
 export default async function createGoodsSpu({ tag, spu: draft }) {
   try {
-    // 1. 创建spu
-    const id = await goodsModel.spuCreate({
-      tag,
-      saasId: saasId(),
-      cId: draft.cId,
-      title: draft.title,
-      desc: draft.desc || '',
-      supplierId: draft.supplier?._id || '',
-      supplierName: draft.supplier?.name || '',
-    });
-
-    // 2. 需要从store中删除原来的spu，然后重新添加 // fixme 为什么不set ？？
-    runInAction(() => {
+    // 尚未创建spu，新建spu
+    if (draft._id.startsWith('draft-')) {
+      // 1. 创建spu
+      const id = await goodsModel.createSpu({
+        tag,
+        param: {
+          // 分类信息
+          cId: draft.cId,
+          // 供应商信息
+          supplierId: draft.supplier?._id || '',
+          supplierName: draft.supplier?.name || '',
+          // 商品信息
+          title: draft.title,
+          desc: draft.desc || '',
+        },
+      });
+      // 更新草稿信息
+      // 删除原来的spu草稿信息
       goodsStore.deleteGoodsSpu(draft._id);
       // 更新ID,重新添加
       draft._id = id;
-      draft.supplierId = draft.supplier._id;
-      draft.suppliername = draft.supplier.name;
       goodsStore.addGoodsSpu({ tag, spu: draft });
-    });
+    } else {
+      log.info(tag, 'createGoodsSpu', 'draft has already been created');
+    }
     log.info(tag, 'createGoodsSpu success!', draft);
+    return spu;
   } catch (error) {
     log.info(tag, 'createGoodsSpu error!', error);
     throw error;
